requirements.txt
anyio==3.6.2
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
exceptiongroup==1.0.0rc9
geckodriver-autoinstaller==0.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wsproto==1.2.0
